Lesson 7, Level 5, Task 7

+++ src/main/java/students/linda_junkina/lesson_7/level_5/task_7/PowerCalculator.java
@@ -3,7 +3,7 @@ package students.linda_junkina.lesson_7.level_5.task_7;
 class PowerCalculator {
     int raisedToPower(int a, int b) {
         int result = 1;
-        for (double i = 1; i <= b; i++) {
+        for (int i = 1; i <= b; i++) {
             result = result * a;
         }
         return result;
